force the output DQ to to int16. The literal 64 defaults to an int64
git-svn-id: http://chara.hi.gemini.edu/svn/DRSoftware/gemini_python/trunk@4430 e04282df-fb77-4b70-bf7a-c1223586725c

+++ astrodata_Gemini/RECIPES_Gemini/primitives/primitives_preprocess.py
@@ -735,7 +735,7 @@ class PreprocessPrimitives(GENERALPrimitives):
                 unilum = np.where(
                         (sci_data>upper) | (sci_data<lower), 64, 0)
 
-                dq_data = np.bitwise_or(dq_data,unilum)
+                dq_data = np.bitwise_or(dq_data,unilum).astype(np.int16)
 
                 # Now replace the DQ data
                 ad[DQ,extver].data = dq_data
